Move `UnionFind` into `UnionFind.Internal`
